orchestrator: intergrate LSM information with device ls
libstoragemgmt metatdata is now available with ceph-volume,
so this patch reformats the plain text output mode of the
orch device ls command to include device health and LED
states.

In addition ; the headings have been camel cased for readability,
the device size is given in dec (since you buy drives in GB not GiB),
a new wide mode has been added and the default output focuses
on the most frequently referred to fields to keep it brief.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>