pybind/mgr/mgr_module: make kwargs explicit
Instead of assuming a kwarg when a '=' is encountered, this commit
changes the CLICommand decorator to require an explicit 'kw=true'
specification in a command definition to add the option to parse it
as a kwarg.
Arguments marked with kw=true can either be passed kwarg style
(arg_name=value) or as a positional argument. Once a kwarg style
argument was seen, the following ones must not be passed as positionals.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>