mgr/cephadm: updates to determining version of crash
The crash daemon is not a real ceph daemon (it's not a client or service),
so instead of running a ceph -version in the creash container we pick up
the version from the environment variables of the container. The version for a crash daemon is defined as master, octopus etc.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>