mon/AuthMonitor: add option to skip non-mon cap parsing
The monitor is not always the same version as the daemons, so it may
not always treat the caps the same. This parsing is also quite cpu
intensive, and is just a usability feature here - it's checked again
on the relevant daemon when the entity connects. Thus, provide an
option to disable this parsing. Leave it on by default for better
usability.

Signed-off-by: Josh Durgin <jdurgin@redhat.com>