mon/OSDMonitor: only take in osd into consideration when trimming osdmaps
we should not take down osd into consideration when trimming osdmap. in
e62269c892, we decrease the upper bound of range of osdmaps to be trimmed
if the given osd is out. but we should have to decrease it only if the
osd in question is still *in*.

so, in this change, the min_lec is decreased only if the osd in question
is *in*.

Fixes: https://tracker.ceph.com/issues/47290
Signed-off-by: Kefu Chai <kchai@redhat.com>