rgw: implement RGWRados::with_bilog<>(), part 2.
This commit is focused on switching to `bi_updater` the part
of `RGWRados::cls_obj_complete_op()` that is responsible
for handling AIO completions. It's necessary as `bi_updater`
is the only component responsible for orchestrating BILog
updates. At the moment this boils down to just:

  `svc.zone->get_zone().log_data`

but will be extended by further commits to dispatch between
the cls_fifo-based and omap-based BILog's backends.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>