common/admin_socket: always validate the parameters
this change addresses the regression introduced by
65267d55cd3714c436b188aaa8b2049ad2b21225.

before this change, we reply the client with a failure only if
both the prefix and the arguments do not match.

after this change, both the prefix and argument are checked.

Fixes: https://tracker.ceph.com/issues/47179
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 0039eb2ce612fff21be71254a33883df80a934b7)