test: remove a now-erroneous "mon add" CLI test from ceph_test_argparse.py
The "mon add" command now lets you pass in arbitrary numbers of strings,
so that you can include locations, so this test is invalid.

I considered updating it to only allow a single non-spaced string, but

datacenter=site1 rack=abc host=host1

is accepted elsewhere, so let's keep that consistent and just remove
this test instead.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>