crimson/osd: add_recovery(oid) before recover_object(oid)
in 7e5ab8811ad64bfe3cd0d00f9136361afd29b8f8, we assumed that the object
to be recovered did not exist in `recovering` before
`recover_object(oid)` was called. but this turns out not true. so, in
this change, `add_object(oid)` is called before `recover_object(oid)`
gets called.

Fixes: https://tracker.ceph.com/issues/47593
Signed-off-by: Kefu Chai <kchai@redhat.com>