mgr/cephadm: get rbd-mirror daemon-id when checking for strays
Currently, list_servers() gets the rbd-mirror service-id instead
of the daemon-id so the daemon is marked as stray. This PR uses
that service-id to find the daemon-id and uses that to check if
the daemon is stray.

Fixes: https://tracker.ceph.com/issues/47639

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 83d878f3d94a31c03ce94788236bd584c878a89c)