qa: Move fs creation/removal to CephFSTestCase class
There is no functional change with this patch. It's
re-org of the code.

The 'CephFSTestCase' class creates the filesystem based on
'REQUIRE_FILESYSTEM' flag in setUp. The 'TestVolume' class
which inherits 'CephFSTestCase' also has the code to create
filesystem which never gets executed. This patch removes this
redundant and dead code.

Also the 'CephFSTestCase' class doesn't cleanup the filesystem
created in setUp in the corresponding tearDown function but
instead cleans up any existing filesystems in setUp prior to
new filesystem creation. Though there is no functional impact,
cleaning it up in tearDown is neat. This patch does the same.
It also retains the filesysytem clean up code in setUp as is
which helps running the test suit on cluster with existing
filesystems.

Fixes: https://tracker.ceph.com/issues/46769
Signed-off-by: Kotresh HR <khiremat@redhat.com>