librbd: fix double decrypt bug when data is read from parent
The crypto object dispatch layer currently always decrypts the data read from the core layer.
However, if the data is actually read from parent, or is all zeros the data should not be decrypted.
This is exactly the case where the child object does not exist.
This commit adds a new flag to mark this case and allow the crypto layer to skip decryption accordingly.

Signed-off-by: Or Ozeri <oro@il.ibm.com>