rgw/kms: Replace boost::string_view with std::string_view
There is a similar commit in master
268f75bb8d3 rgw: Replace boost::string_ref/view with std::string_view
but this commit replaces string_view in many more places pieces than
just kms, and did not apply cleanly to nautilus.

Instead, this commit affects *just* kms, (which itself is not native to
nautilus), and provides a few adapter inlines so that the remainder of
nautilus need not change.

Signed-off-by: Marcus Watts <mwatts@redhat.com>