cephadm: Push ceph-iscsi config options into mon-store
Currently the ceph-iscsi config options get placed in the
iscsi-gateway.cfg but this is volume mounted into the container and
lives under `/var/lib/ceph/<fsid>/<iscsi service>/iscsi-gateway.cfg`. This is
fine for day 1 situations but horrible on day 2.

To make this easier a PR was pushed to ceph-iscsi[0] which taught it how
to look in the mon config-key store for values to config options. This
took the same approach as RGW. A config value starting with a
`config://` will be treated as key in the mon store. E.g:

  config://my/awesome/key

Will go look for the value of `my/awesome/key`.

Now that the ceph-iscsi PR has merged, it's time to put this into
cephadm/orch.
This patch adds a new option to the iscsi spec called `mon_key_prefix`
which defaults to `iscsi`. Which allows someone to deploy a ceph-iscsi
service with either globally shared options or can seperate as required.

Note that because we set config-key get cap for the iscsi user to only
access keys starting with `iscsi/` this is also a hard requirement. By
default it'll place config options under `iscsi/..` keys:

  iscsi/trusted_ip_list
  iscsi/api_port
  iscsi/api_user
  iscsi/api_password
  iscsi/api_secure

If you want to have say 2 different sets some these options you could
always deploy services with different `mon_key_prefix`'s say as
`iscsi/datacenterA` and `iscsi/datacenterB`. Then they look at different
locations:

  iscsi/datacenterA/trusted_ip_list
  iscsi/dataventerA/api_port
  ...
  iscsi/datacenterB/trusted_ip_list
  iscsi/dataventerB/api_port
  ...

Best part is, now if you want to change a value, you don't need to dig
around and find a config file or do a redeloy, just run a `ceph config-key set` and
then restart the service.

[0] https://github.com/ceph/ceph-iscsi/pull/217

Signed-off-by: Matthew Oliver <moliver@suse.com>