doc/cephfs: updated top-level links
Top-level links weren't working, so I collected
them into .rst pages that act like target nodes,
and now the structure of the book is a bit more
evident and foregrounded than it was. It's
quite possible that we might want to alter this
arrangement of the TOC so that it exposes less
stuff, but at the very least this fixes the
problem of links going nowhere.

Fixes: https://tracker.ceph.com/issues/45344
Signed-off-by: Zac Dover <zac.dover@gmail.com>