mds: ignore sequence values on REQUEST_CLOSE
When unmounting, clients send a REQUEST_CLOSE and just wait for the
reply before eventually giving up after a timeout and proceeding. If the
request crossed paths with a message from the server that bumped the seq
number, then the MDS will drop the request on the floor.

The clients never retransmit this request so the client just hangs,
sometimes even resulting in a blocklisting event when caps aren't
returned.

A REQUEST_CLOSE is an indicator that the client is unmounting. It's not
going to decide to not do that because the sequence number changed. Make
the MDS ignore the sequence on a REQUEST_CLOSE.

Fixes: https://tracker.ceph.com/issues/47563
Signed-off-by: Jeff Layton <jlayton@redhat.com>