cephadm: mounted directory in rbd-mirror for remote cluster info
When you add a peer with rbd mirror it searches /etc/ceph for
the conf file and keyring for the remote directory. Since cephadm
deploys the service in a container the user would have to copy
those files into the container for it to work which means they
would have to redo this every time an rbd-mirror container is
redeployed. This provides a directory on the host machine that
is mounted to the rbd-mirror container the user can put those
files in instead that will be persistant through redeploys

Signed-off-by: Adam King <adking@redhat.com>