qa/suites/rados/singleton-nomsgr: add diskprediction_local integration tests
In order to verify that diskpredicion_local models function properly
across all supported distribution, we can simply run their unit tests
via this rados suite.
We do not need to bring up a cluster since we use smartctl JSON samples
to mock the input to the models.

This by itself will mock the loading of the required dependencies, and
will provide us with an accurate indication regarding the behavior of
diskprediction_local module on each distribution.

Please note:
1. These tests will fail when:
  - sklearn or numpy are not installed.
  - sklearn or numpy are installed, but with different (older or newer)
    versions which cause the code to break (due to lack of backward
    compatibility, or API changes).

2. Currently the tests print warnings that several functions used by the
   models will be deprecated in newer sklearn versions, but the tests are
   considered to pass successfully.

3. Nautilus runs python 2. sklearn 0.20 was the last version to support
   Python 2.7.

Fixes: https://tracker.ceph.com/issues/47448
Signed-off-by: Yaarit Hatuka <yaarit@redhat.com>