rgw/gc: fixing the condition when marker for a queue is always reset to empty which causes RGWGC::list to get stuck in a loop, which ultimately is broken out of when the queue's truncated flag is false.
1. Check for entries size also while evaluating whether objects cache for
a gc object should be marked as 'transitioned' in case of cls_rgw_gc_list.
When there are no entries, we get back a return value of 0, and the
object cache is not marked as 'transitioned'.

2. Also for the last gc object, we need to check whether the queue is still
under process and set the correct flag.

Missing the two conditions above causes the GC::list to loop continously
over the same gc object.

Fixes: https://tracker.ceph.com/issues/47909

Signed-off-by: Pritha Srivastava <prsrivas@redhat.com>