RGW: Fix Swift GET container metadata
Currently when you run a GET on a container the X-Container-* metadata
returned is zeroed out and doesn't actaully return the current stats for
the container. If you do a HEAD however it behaves as expected.

This is because on a HEAD the container object is actaully pulled from
the store. So this patch does the same thing for a GET.

I pushed this back into the RGWListBucket class, so should benefit S3
and Swift apis, however if poeple have a problem with this, we could
just do it in the Swift child class, RGWListBucket_ObjStore_SWIFT.

Fixes: https://tracker.ceph.com/issues/47861
Fixes: https://tracker.ceph.com/issues/47781
Signed-off-by: Matthew Oliver <moliver@suse.com>