crimson/mon: actually skip CEPHX_V2 authorizer challenges when needed
When commit 2c1b133e5987 ("mon/MonClient: skip CEPHX_V2 challenge
if client doesn't support it") was being ported to crimson, it got
adjusted incorrectly (but rather fortunately).  While the original
always skips authorizer challenges, the port almost never skips
them.  This is because it looks at the features of the monitor that
crimson-osd is connected to instead of the features of the peer that
is trying to authenticate.  As a result, msgr1 peer would get the
challenge even if it legitimately doesn't support CEPHX_V2 and fault
on an unknown protocol tag.

To fix, port commit 95061b322f80 ("mon/MonClient: bring back CEPHX_V2
authorizer challenges").

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>