rgw: delegate the max_marker handling to RGWSI_BILog_RADOS.
In this commit `get_log_status()` of `RGWSI_BILog_RADOS` gets
replaced by `lot_get_max_marker()`. The main difference between
them, apart naming (which better reflects the purpose, I hope),
is the arrangement of input. After the rework it's a caller's
responsibility to provide the `rgw_bucket_dir_header` map.
That is, the method is no longer tasked with retrieving it which
was accomplished via `RGWSI_BucketIndex_RADOS::cls_bucket_head()`.
The motivation here is the need to differentiate the `max_marker`
handling between the BILog backends.

The commit also renames the `cls_bucket_head()` methods to
`get_dir_headers()`.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>