RGW: Check all shards for user manifest parts
When we go and read an user manifest LO (or a DLO in the swift api)
there is a bug where we supply a defaulted params object that defaults
shard_id to 0. This means that when issueing a GET to the LO it'll only
check shard 0 for parts.

This is causing errors in tempest tests, and is obviously wrong.

This patch defaults the ListParams.shard_id to -1, so by default
all shards are checked for parts and therefore complete LO's
are returned.

Fixes: https://tracker.ceph.com/issues/47801
Fixes: https://tracker.ceph.com/issues/47800
Signed-off-by: Matthew Oliver <moliver@suse.com>