Don't writeback when inside a readdir callback
The code updates some attributes that may have changed between versions
when doing a lookup.  However, a lookup can be done from within a
readdir callback, which causes a deadlock during the writeback.  Extend
the flag indicating we're in a callback all the way down, so we can
conditionally avoid the write.

Fixes RHBZ#1845501

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>