librbd: relax requirements on finisher canceled callback
The finisher timer is started with safe_callbacks = false, and
cancel_event may fail.

When canceling a task it is safe to just ignore the cancel_event
result and proceed, because the returned false value means the
callback is in TaskFinisher::complete already but before
acquiring the lock, so when it eventually acquires the lock it
will just find out the task is already deleted and return.

Signed-off-by: Mykola Golub <mgolub@suse.com>